requirements.txt
streamlit
pandas
requests
matplotlib
boto3
python-dotenv
pytz